requirements.txt
astroid==2.3.3
autopep8==1.4.4
Click==7.0
colorama==0.4.3
emojis==0.5.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
prompt-toolkit==1.0.14
pycodestyle==2.5.0
pyfiglet==0.8.post1
Pygments==2.5.2
PyInquirer==1.0.3
pylint==2.4.4
PyYAML==5.3
regex==2020.1.8
six==1.13.0
typed-ast==1.4.0
wcwidth==0.1.8
wrapt==1.11.2
